Create FINANCIAL ENGINEERING IN INTEREST RATES & FX - CITY UNIVERSITY. LONDON
Price on Binomial Model, Trinomial Model, and  Monte Carlo models European Options, Digitals, Spread Options. Then extend our implementation to Barriers, Double-No-Touches. On the American Optionality, we will look at the switch Optionality for American Options on the trees and then on to the Monte-Carlo we will investigate the Longstaff-Schwartz methodology for exercise decision strategy on multi-call options. Then we will transition into pricing implementation for Swaptions and variance Reduction techniques on Sum of Libor Options.